+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - fixed overflowing of the code blocks in messages
 - pasting multiple MCP servers at once within the same JSON will add all the servers
 - revamped session management
+- adding command output to context messages for Agent
 
 ## [0.8.0]